Transform car entity to schema
